+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG for PHP CS Fixer: custom fixers
 
+## [Unreleased]
+- Add SingleLineThrowFixer
+
 ## v1.13.0 - *2019-01-11*
 - Update PHP CS Fixer to v2.14
 - OperatorLinebreakFixer - respect no whitespace around operator